+++ tests/test_obj_BaseGeo.py
