+++ tests/test_obj_Box.py
@@ -1,15 +1,17 @@
-import os
 import pickle
+import os
 import numpy as np
 import magpylib as mag3
-from magpylib.magnet import Cylinder
+from magpylib.magnet import Box
+from magpylib._lib.obj_classes.class_Sensor import Sensor
 
-# # GENERATE DATA
-# N = 22
+# """data generation for test_Box()"""
+
+# N = 100
 
 # mags = (np.random.rand(N,3)-0.5)*1000
-# dims = (np.random.rand(N,2)-0.5)*5
-# posos = (np.random.rand(N,77,3)-0.5)*10 #readout at 333 positions
+# dims = (np.random.rand(N,3)-0.5)*5
+# posos = (np.random.rand(N,333,3)-0.5)*10 #readout at 333 positions
 
 # angs =  (np.random.rand(N,18)-0.5)*2*10 # each step rote by max 10 deg
 # axs =   (np.random.rand(N,18,3)-0.5)
@@ -18,50 +20,55 @@ from magpylib.magnet import Cylinder
 
 # B = []
 # for mag,dim,ang,ax,anch,mov,poso in zip(mags,dims,angs,axs,anchs,movs,posos):
-#     pm = Cylinder(mag,dim)
+#     pm = Box(mag,dim)
+
 #     # 18 subsequent operations
 #     for a,aa,aaa,mv in zip(ang,ax,anch,mov):
 #         pm.move_by(mv).rotate_from_angax(a,aa,aaa)
-#     B += [pm.getB(poso,niter=100)]
+
+#     B += [pm.getB(poso)]
 # B = np.array(B)
+
 # inp = [mags,dims,posos,angs,axs,anchs,movs,B]
-# pickle.dump(inp,open('testdata_Cylinder.p', 'wb'))
+
+# pickle.dump(inp,open(os.path.abspath('./../'testdata_Box.p', 'wb'))
 
 
-def test_Cylinder_basics():
-    """  test Cylinder fundamentals, test against magpylib2 fields
+def test_Box_basics():
+    """ test Box fundamentals, test against magpylib2 fields
     """
-    data = pickle.load(open(os.path.abspath('./tests/testdata/testdata_Cylinder.p'), 'rb'))
+    # data generated below
+    data = pickle.load(open(os.path.abspath('./tests/testdata/testdata_Box.p'), 'rb'))
     mags,dims,posos,angs,axs,anchs,movs,B = data
 
-    Btest = []
+    btest = []
     for mag,dim,ang,ax,anch,mov,poso in zip(mags,dims,angs,axs,anchs,movs,posos):
-        pm = Cylinder(mag,dim)
+        pm = Box(mag,dim)
 
         # 18 subsequent operations
         for a,aa,aaa,mv in zip(ang,ax,anch,mov):
             pm.move_by(mv).rotate_from_angax(a,aa,aaa)
 
-        Btest += [pm.getB(poso, niter=100)]
-    Btest = np.array(Btest)
+        btest += [pm.getB(poso)]
+    btest = np.array(btest)
 
-    assert np.allclose(B, Btest), "test_Cylinder failed big time"
+    assert np.allclose(B, btest), "test_Box failed big time"
 
 
-def test_Cylinder_add():
+def test_Box_add():
     """ testing __add__
     """
-    src1 = Cylinder((1,2,3),(1,2))
-    src2 = Cylinder((1,2,3),(1,2))
+    src1 = Box((1,2,3),(1,2,3))
+    src2 = Box((1,2,3),(1,2,3))
     col = src1 + src2
-    assert isinstance(col,mag3.Collection), 'adding cylinder fail'
+    assert isinstance(col,mag3.Collection), 'adding boxes fail'
 
 
-def test_Cylinder_squeeze():
+def test_Box_squeeze():
     """ testing squeeze output
     """
-    src1 = Cylinder((1,1,1),(1,1))
-    sensor = mag3.Sensor(pos_pix=[(1,2,3),(1,2,3)])
+    src1 = Box((1,1,1),(1,1,1))
+    sensor = Sensor(pos_pix=[(1,2,3),(1,2,3)])
     B = src1.getB(sensor)
     assert B.shape==(2,3)
     H = src1.getH(sensor)
@@ -71,3 +78,10 @@ def test_Cylinder_squeeze():
     assert B.shape==(1,1,1,2,3)
     H = src1.getH(sensor,squeeze=False)
     assert H.shape==(1,1,1,2,3)
+
+
+def test_repr_box():
+    """ test __repr__
+    """
+    pm1 = Box((1,2,3),(1,2,3))
+    assert pm1.__repr__()[:3] == 'Box', 'Box repr failed'
